known aliases for currency, from openexchangerates.org
fyi they have others, which they call "black market". notably in the black market: several cryptocurrencies. Notably *not* in the black market: bitcoin. Also notably not in the black market: several fantasies of the financial market. So if it's the legal argument, i don't get why bitcoin gets privilege. And obviously it's not the moral argument.

oh also a few jokes from myself.

see #28